+++ feature/savelink/src/main/java/org/sopt/savelink/ui/model/Clip.kt
@@ -5,7 +5,7 @@ import org.sopt.model.category.Category
 data class Clip(
   val categoryId: Long?,
   val categoryTitle: String,
-  val toastNum: Int,
+  val toastNum: Long,
   var isSelected: Boolean,
 )
